requirements.txt
boto3==1.10.21
botocore==1.13.21
dj-database-url==0.5.0
Django==2.2.7
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-storages==1.7.2
docutils==0.15.2
gunicorn==20.0.0
jmespath==0.9.4
Pillow==6.2.1
psycopg2==2.8.4
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
django-star-ratings==0.9.1